# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Interview Stack Guide API
  version: 1.1.0
  description: REST API reference for interview preparation — dual API with GraphQL at /graphql
servers:
  - url: http://localhost:4000
paths:
...
  /api/products:
    get:
      summary: List products (paginated)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: cursor
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: inStock
          in: query
          schema: { type: string, enum: [true, false] }
      responses:
        "200":
          description: Paginated product list
    post:
      summary: Create product (auth required when ENABLE_AUTH=true)
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /api/products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product found
        "404":
          description: Not found
    put:
      summary: Update product
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete product
      responses:
        "204":
          description: Deleted
...